Remove Unnecessary Workers
