--- app/src/main/java/com/latticeonfhir/android/service/workmanager/workers/download/patient/PatientDownloadSyncWorker.kt
@@ -1,41 +1,0 @@
-package com.latticeonfhir.android.service.workmanager.workers.download.patient
-
-import android.content.Context
-import androidx.work.WorkerParameters
-import androidx.work.workDataOf
-import com.latticeonfhir.android.service.workmanager.workers.base.SyncWorker
-import com.latticeonfhir.android.utils.constants.ErrorConstants
-import com.latticeonfhir.android.utils.converters.server.responsemapper.ApiContinueResponse
-import com.latticeonfhir.android.utils.converters.server.responsemapper.ApiEmptyResponse
-import com.latticeonfhir.android.utils.converters.server.responsemapper.ApiEndResponse
-import com.latticeonfhir.android.utils.converters.server.responsemapper.ApiErrorResponse
-
-abstract class PatientDownloadSyncWorker(context: Context, workerParameters: WorkerParameters) :
-    SyncWorker(context, workerParameters) {
-
-    override suspend fun doWork(): Result {
-        setProgress(workDataOf(PatientDownloadProgress to 0))
-
-        return when (val response = getSyncRepository().getAndInsertListPatientData(0)) {
-            is ApiContinueResponse -> Result.success()
-            is ApiEndResponse -> {
-                setProgress(workDataOf(PatientDownloadProgress to 100))
-                Result.success()
-            }
-
-            is ApiEmptyResponse -> Result.failure()
-            is ApiErrorResponse -> {
-                if (response.errorMessage == ErrorConstants.SESSION_EXPIRED || response.errorMessage == ErrorConstants.UNAUTHORIZED) Result.failure(
-                    workDataOf("errorMsg" to response.errorMessage)
-                )
-                else Result.retry()
-            }
-
-            else -> Result.retry()
-        }
-    }
-
-    companion object {
-        const val PatientDownloadProgress = "PatientDownloadProgress"
-    }
-}
