--- app/src/main/java/com/latticeonfhir/android/service/workmanager/workers/download/patient/PatientDownloadSyncWorkerImpl.kt
@@ -1,10 +1,0 @@
-package com.latticeonfhir.android.service.workmanager.workers.download.patient
-
-import android.content.Context
-import androidx.work.WorkerParameters
-import com.latticeonfhir.android.FhirApp
-
-class PatientDownloadSyncWorkerImpl(context: Context, workerParameters: WorkerParameters) :
-    PatientDownloadSyncWorker(context, workerParameters) {
-    override fun getSyncRepository() = (applicationContext as FhirApp).getSyncRepository()
-}
