--- app/src/main/java/com/latticeonfhir/android/service/workmanager/workers/upload/patient/patch/PatientPatchUploadSyncWorkerImpl.kt
@@ -1,11 +1,0 @@
-package com.latticeonfhir.android.service.workmanager.workers.upload.patient.patch
-
-import android.content.Context
-import androidx.work.WorkerParameters
-import com.latticeonfhir.android.FhirApp
-
-class PatientPatchUploadSyncWorkerImpl(context: Context, workerParameters: WorkerParameters) :
-    PatientPatchUploadSyncWorker(context, workerParameters) {
-
-    override fun getSyncRepository() = (applicationContext as FhirApp).getSyncRepository()
-}
