--- app/src/main/java/com/latticeonfhir/android/service/workmanager/workers/upload/prescription/PrescriptionUploadSyncWorkerImpl.kt
@@ -1,10 +1,0 @@
-package com.latticeonfhir.android.service.workmanager.workers.upload.prescription
-
-import android.content.Context
-import androidx.work.WorkerParameters
-import com.latticeonfhir.android.FhirApp
-
-class PrescriptionUploadSyncWorkerImpl(context: Context, workerParameters: WorkerParameters) :
-    PrescriptionUploadSyncWorker(context, workerParameters) {
-    override fun getSyncRepository() = (applicationContext as FhirApp).getSyncRepository()
-}
